Oil, gears and scripts
Replaced old red placeholder gears with (static) oil sprite (needs to be
animated!). Also, it looks like that player will die before it collides
with oil.
Added some bg gears, and related to this, I tweaked (copies of) the
rotation script to make them rotate in a sync (though I'm not using
exact values there so in time they will be out of sync but hopefully
no-one will stare at them for hours...)